Align the columns of a pander table.

Uses pander() to print a data frame as a table to the output document.

align_pander(x, align_idx = NULL, caption = NULL)

Arguments

x
A data frame to be printed in the output document.

align_idx
Optional string made up of l (left-aligned), r (right-aligned), and c (center-aligned).

caption
Optional string used as the pander() caption argument.

Value

Prints the data frame in table form using pander(x).

Details

Uses panderOptions('table.alignment.default') argument to assign output column alignment. The default alignments are numeric right and everything else left.

If align_idx is not NULL, a warning message is generated if its string length does not match the number of columns in the table.

The function also sets these pander options as defaults:

table.split.table = Inf

keep.trailing.zeros = TRUE

style = 'simple'

Examples

x <- mtcars[1:5, 1:5]
align_pander(x)

#> 
#> 
#>         &nbsp;             mpg   cyl   disp    hp   drat
#> ----------------------- ------ ----- ------ ----- ------
#>      **Mazda RX4**        21.0     6    160   110   3.90
#>    **Mazda RX4 Wag**      21.0     6    160   110   3.90
#>     **Datsun 710**        22.8     4    108    93   3.85
#>   **Hornet 4 Drive**      21.4     6    258   110   3.08
#>  **Hornet Sportabout**    18.7     8    360   175   3.15
#> 

align_pander(x, align_idx = "rcrrr")

#> 
#> 
#>         &nbsp;             mpg  cyl    disp    hp   drat
#> ----------------------- ------ ----- ------ ----- ------
#>      **Mazda RX4**        21.0   6      160   110   3.90
#>    **Mazda RX4 Wag**      21.0   6      160   110   3.90
#>     **Datsun 710**        22.8   4      108    93   3.85
#>   **Hornet 4 Drive**      21.4   6      258   110   3.08
#>  **Hornet Sportabout**    18.7   8      360   175   3.15
#> 

align_pander(x, caption = "A nicely formatted table")

#> 
#> 
#>         &nbsp;             mpg   cyl   disp    hp   drat
#> ----------------------- ------ ----- ------ ----- ------
#>      **Mazda RX4**        21.0     6    160   110   3.90
#>    **Mazda RX4 Wag**      21.0     6    160   110   3.90
#>     **Datsun 710**        22.8     4    108    93   3.85
#>   **Hornet 4 Drive**      21.4     6    258   110   3.08
#>  **Hornet Sportabout**    18.7     8    360   175   3.15
#> 
#> Table: A nicely formatted table
#>