Deprecated functions in package docxtools

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("docxtools-deprecated").

align_pander(x, align_idx = NULL, caption = NULL)

align_pander

For align_pander, use kable::knitr.