docxtools: R Markdown to docx helper functions.

The docxtools package provides a set of helper functions for using R Markdown to create documents in docx format, especially documents for use in a classroom or workshop setting.

Details

The package provides two categories of functions: one for formatting numbers and tables; and one for placing specific objects in a docx output document.

Number and table formatting

format_engr() Format numerical variables in engineering notation. align_pander() Align the columns of a pander table.

Insert objects into docx documents

put_axes() Insert a two-dimensional coordinate axes in a document. put_gap() Insert a gap or whitespace in a document.