Format numerical variables in engineering notation.

Converts numeric objects into character variables formatted in engineering notation: base 10 with exponents that are multiples of 3. The input argument is a numeric vector or a data frame with at least one numeric variable.

format_engr(x, sigdig = NULL)

Arguments

x
: The numeric object to be formatted.

sigdig
: An optional vector of significant digits.

Value

The returned object is a data frame with numeric variables converted to formatted strings bounded by "$...$" and interpreted by R Markdown as an equation.

Details

The preferred input is a data frame. If the input is a numeric vector, it will be formatted and returned as a data frame with the variable name value. To preserve the variable name from the calling script, convert the number or vector to a data frame before calling engr_format().

Non-numeric variables in a data frame are unaffected and are returned as-is.

Each numeric variable can be assigned its own number of significant digits. The default is 4. If sigdig is a single value, it is used for all variables; if an array, an excess number of elements is truncated and a shortage of elements is filled with the default 4. Setting sigdig = 0 returns the numerical variable as-is.

The output format is "coefficient x 10^exponent" except for exponents equal to 0, 1, or 2. In these cases, the output format is floating-point. In all cases, the decimal point in the coefficient floats.

Trailing zeros to the right of a decimal are significant. A trailing zero before a decimal is ambiguous; such numbers are reported with the decimal moved an additional three places to the left except for cases in which doing so results in a zero coefficient due to a small number of significant digits.

The numerical output strings are bounded by "$...$" for rendering in an R markdown output document, e.g., HTML, PDF, or DOCX. For example, the Rmd code chunk knitr::kable(engr_format(df)) prints the formatted data frame to the output document.

To learn more about docxtools, start with the vignettes: browseVignettes(package = "docxtools").

Examples

# create test input arguments
set.seed(20161221)
n <- 5
a <- sample(letters, n)
b <- sample(letters, n)
w <- runif(n, min =  -5, max =  50) * 1e+5
y <- runif(n, min = -25, max =  40) / 1e+3
z <- runif(n, min =  -5, max = 100)
x <- data.frame(z, b, y, a, w, stringsAsFactors = FALSE)

# format different objects
print(x)

#>           z b            y a         w
#> 1  6.501440 c  0.001051893 q 2846529.3
#> 2 28.374092 o  0.000347614 y 4874357.1
#> 3 -3.849624 i  0.004599897 g -111651.4
#> 4 44.500979 a -0.003045062 a 1314715.7
#> 5 92.411835 x -0.001069473 i  417385.0

format_engr(x)

#> Warning: the condition has length > 1 and only the first element will be used

#>          z b                         y a                        w
#> 1  $6.501$ c  ${1.052}\\times 10^{-3}$ q  ${2.847}\\times 10^{6}$
#> 2  $28.37$ o  ${347.6}\\times 10^{-6}$ y  ${4.874}\\times 10^{6}$
#> 3 $-3.850$ i  ${4.600}\\times 10^{-3}$ g ${-111.7}\\times 10^{3}$
#> 4  $44.50$ a ${-3.045}\\times 10^{-3}$ a  ${1.315}\\times 10^{6}$
#> 5  $92.41$ x ${-1.069}\\times 10^{-3}$ i  ${417.4}\\times 10^{3}$

format_engr(x, sigdig = 3)

#> Warning: the condition has length > 1 and only the first element will be used

#>         z b                        y a                      w
#> 1  $6.50$ c  ${1.05}\\times 10^{-3}$ q ${2.85}\\times 10^{6}$
#> 2  $28.4$ o   ${348}\\times 10^{-6}$ y ${4.87}\\times 10^{6}$
#> 3 $-3.85$ i  ${4.60}\\times 10^{-3}$ g ${-112}\\times 10^{3}$
#> 4  $44.5$ a ${-3.05}\\times 10^{-3}$ a ${1.31}\\times 10^{6}$
#> 5  $92.4$ x ${-1.07}\\times 10^{-3}$ i  ${417}\\times 10^{3}$

format_engr(x, sigdig = c(3, 4, 5))

#> Warning: the condition has length > 1 and only the first element will be used

#>         z b                         y a                         w
#> 1  $6.50$ c  ${1.052}\\times 10^{-3}$ q  ${2.8465}\\times 10^{6}$
#> 2  $28.4$ o  ${347.6}\\times 10^{-6}$ y  ${4.8744}\\times 10^{6}$
#> 3 $-3.85$ i  ${4.600}\\times 10^{-3}$ g ${-111.65}\\times 10^{3}$
#> 4  $44.5$ a ${-3.045}\\times 10^{-3}$ a  ${1.3147}\\times 10^{6}$
#> 5  $92.4$ x ${-1.069}\\times 10^{-3}$ i  ${417.38}\\times 10^{3}$

format_engr(y, 2)  # numeric vector

#> Warning: the condition has length > 1 and only the first element will be used

#>                     value
#> 1  ${1.1}\\times 10^{-3}$
#> 2 ${0.35}\\times 10^{-3}$
#> 3  ${4.6}\\times 10^{-3}$
#> 4 ${-3.0}\\times 10^{-3}$
#> 5 ${-1.1}\\times 10^{-3}$

format_engr(n, 3)  # numeric scalar

#>    value
#> 1 $5.00$

# using base R data sets
format_engr(DNase[1:10, ], sigdig = 4)

#> Warning: the condition has length > 1 and only the first element will be used

#>    Run                     conc                  density
#> 1    1 ${48.83}\\times 10^{-3}$ ${17.00}\\times 10^{-3}$
#> 2    1 ${48.83}\\times 10^{-3}$ ${18.00}\\times 10^{-3}$
#> 3    1 ${195.3}\\times 10^{-3}$ ${121.0}\\times 10^{-3}$
#> 4    1 ${195.3}\\times 10^{-3}$ ${124.0}\\times 10^{-3}$
#> 5    1 ${390.6}\\times 10^{-3}$ ${206.0}\\times 10^{-3}$
#> 6    1 ${390.6}\\times 10^{-3}$ ${215.0}\\times 10^{-3}$
#> 7    1 ${781.2}\\times 10^{-3}$ ${377.0}\\times 10^{-3}$
#> 8    1 ${781.2}\\times 10^{-3}$ ${374.0}\\times 10^{-3}$
#> 9    1                  $1.562$ ${614.0}\\times 10^{-3}$
#> 10   1                  $1.562$ ${609.0}\\times 10^{-3}$

format_engr(mtcars[1:5, 1:5], sigdig = c(3, 0, 0, 0, 3))

#> Warning: the condition has length > 1 and only the first element will be used

#>      mpg cyl  disp    hp   drat
#> 1 $21.0$ $6$ $160$ $110$ $3.90$
#> 2 $21.0$ $6$ $160$ $110$ $3.90$
#> 3 $22.8$ $4$ $108$  $93$ $3.85$
#> 4 $21.4$ $6$ $258$ $110$ $3.08$
#> 5 $18.7$ $8$ $360$ $175$ $3.15$

format_engr(CO2[1:5,], 3)

#> Warning: the condition has length > 1 and only the first element will be used

#>   Plant   Type  Treatment                    conc uptake
#> 1   Qn1 Quebec nonchilled                  $95.0$ $16.0$
#> 2   Qn1 Quebec nonchilled                   $175$ $30.4$
#> 3   Qn1 Quebec nonchilled ${0.250}\\times 10^{3}$ $34.8$
#> 4   Qn1 Quebec nonchilled ${0.350}\\times 10^{3}$ $37.2$
#> 5   Qn1 Quebec nonchilled ${0.500}\\times 10^{3}$ $35.3$

# NOT RUN {
# if no columns are numeric, an error is thrown, e.g.,
a
format_engr(a)
# }